Fix isNotRecurringNestStruct function
Currently, the function does not use its refTypeName parameter;
instead it checks that the 1st structure in the structures stack
does not repeat. Recursion, however, can be indirect. Consider the
following example:

    type Foo struct {
        Bar Bar
    }

    type Bar struct {
        Foo Foo
    }

    type Baz struct {
        Foo Foo
    }

In the example generator will crash with stack overflow while trying to
generate spec for Baz when there is no spec for Foo and Bar generated
yet.

The fix is to check that refTypeName does not repeat in the structures
stack.