Allow different names for swagger specs
In the case where we need to include multiple docs packages in the same
process, swag.Register will fail. Despite looking like having support
for swagger specs with different names it actually does not work.

Define RegisterName and ReadDocName functions which allow you to specify
the name of the swag. Default method will still use this path but with
the 'classic' swag name of 'swagger'.