fix: Exclude operation before it's parsed
Previous implementation had an issue where it was excluding operation after schema
and other parts were parsed. That caused schema definition to be included in the
output file, even though operation didn't match the extension.

New implementation won't even start processing operation if extension isn't matching.

Only potential issue/problem I see with this approach is that we are duplicating logic
for comment line parsing (I basically c/p it from other places) but not sure how big
of an issue that actually is as I noticed we are doing that at other places as well.