V2: Update Makefile (#1560)
* Update README_zh-CN.md (#1545)

remove repeat net/http

* Add option to set template delimiters (#1499)

* Add template action delimiter cli flag

* Add delims to generator config and template

Also adds tests using the "quote" test as a base. This has to have a
custom Instance name or it will clash with the "quotes" one and panic
since it will have registered two "swagger" instances in the package
test.

* Add testdata for custom delim flags

Based on the "quote" testdata.

* Add delims to the spec, with tests.

Make sure we don't add delims if they are empty. This shouldn't be
possible, but might as well be safe.

* Go mod tidy and sum update

* Make the CLI experience a bit cleaner

* Revert go.mod and sum

* Update readme

* fix bug: enums of explicit type conversion (#1556)

Signed-off-by: sdghchj <sdghchj@qq.com>

* integrate new delimiter feature from master branch into v2

* cleanup

* update makefile to use v2 path

---------

Signed-off-by: sdghchj <sdghchj@qq.com>
Co-authored-by: tzxdtc10 <tzxdtc10@gmail.com>
Co-authored-by: Leo Palmer Sunmo <leosunmo@users.noreply.github.com>
Co-authored-by: sdghchj <sdghchj@qq.com>
Co-authored-by: Tobias Theel <tt@fino.digital>