Send the version to stderr
Until https://tip.golang.org/doc/modules/managing-dependencies#tools is
available to me, I have a shell script that detects the version of swag
by invoking swag --version and parsing its output, then updates if I'm
not using the version specified in the script. With v2, the output
looks like this

Swag version:  v2.0.0
swag version v2.0.0

While it might be useful to have the swag version always printed for
troubleshooting purposes, I think it should be printed to stderr so as
not to mess up output that is supposed to be parsed.

This change will result in the first line to be sent to stderr, making
it backward-compatible with the previous output of swag --version while
still retaining the improvement of having it printed for every command.