Fixes list recursion (#2022)
* Fixes list recursion by returning any

I had 2 issues.

1) A panic due to nil pointer when
p.openAPI.Components.Spec.Schemas[searchString]
was nil, and getSchemaByRef tried to return .Spec
Resolved with a simple check and nill return (to start)

This highlighted issue 2 which in some ways relates to
https://github.com/swaggo/swag/issues/1939
2) When an ARRAY schema type is evaluated, and getSchemaByRef
returns nil, generation would fail.
So getSchemaByRef is changed to warn and return ANY instead
similar to some other places, which means the generation can
work.

I expect this could be resolved more globally at some point
pulling in other things from the linked issue...

* Fix recursion detection using incorrect name

Prior to this, in cases where recursion was detected,
the annotated @name value was not used in the generated
spec.
This in turn would lead to invalid spec, as some values would
be using the full name, and others the hort name...

* Tests for list recursion fixes