fix: prevent nil pointer dereference in getAllGoFileInfo when parsing file
If there's a file or dir that can't be read, the application throws
a cryptic error that is not easy to debug.
In this fix, we handle the errors properly so that the user knows there
are issues with some file in the path, allowing the app to process all the others.

Fixes: #1916

Signed-off-by: Valério Valério <vdv100@gmail.com>